+++ Connectors/CSi/Speckle.Connectors.CSiShared/ServiceRegistration.cs
@@ -15,7 +15,7 @@ namespace Speckle.Connectors.CSiShared;
 public static class ServiceRegistration
 {
   // TODO: AddCSi and AddETABS for shared and specific implementations respectively. To do with CNX-828
-  public static IServiceCollection AddETABS(this IServiceCollection services)
+  public static IServiceCollection AddCSi(this IServiceCollection services)
   {
     services.AddSingleton<IBrowserBridge, BrowserBridge>();
     services.AddSingleton<ICSiApplicationService, CSiApplicationService>();
