+++ Connectors/CSi/Speckle.Connectors.ETABS22/Speckle.Connectors.ETABS22.csproj
@@ -21,6 +21,12 @@
       <PackageReference Include="Speckle.CSI.API" PrivateAssets="all" IncludeAssets="compile; build" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Compile Update="Form1.cs">
+        <SubType>Form</SubType>
+      </Compile>
+    </ItemGroup>
+
     <Import Project="..\Speckle.Connectors.CSiShared\Speckle.Connectors.CSiShared.projitems" Label="Shared" />
 
 </Project>
